Make it possible (easy?) to compile timel
* Add clean and default target to Makefile
* Use gfortran for all fortran files
* Add libsun and libtau to directory
* Fix libtau for modern fortran:

  NB:  call abort('string') has never been valid
       under any ISO Fortran standard, so replace
       with print*, 'string'; call abort, which is
       at least valid Fortran 90. Also, why would
       one pass a value into the argc() function?

This now "works" and direct can read an ELCOR.dat
file to make elipcore.tbl. But how do we find the
parameters (of the correction sperical harmonics?)
in ELCOR.dat? Is this model dependant? We also need
ak135.tbl and ak135.tbl (or something from other
models) made in the taup directory.