Basic minimum to wrap ellip corr. in python
It turns out that the ellipticiy correction
is (almost) a self-contained function in
Fortran. So with a little change (to move the
location where the formatted file is read
for the correction coefficents) we can turn
this into python extension package.

At this stage I have done the following:

* put ellip.f into a new directory to
  hold the extension module source.

* generated a f2py interface (.pyf) file.
  To do this I started with:
  `f2py -h ellip.pyf -f _ellip_fortran ellip.f`
  then added intents to the arguments (so python
  knows what should be returned and what are input
  arguments). Note that I'm not going to expect
  callers to run the fortran interface so I've
  protected the module name: from memory this will
  only be importable from inside the package.

* added an open statment to ellip.f (so it opens
  the table. This used to be inside ttimel.f.

f2py seems to work well with the (fun) entry
statments (which is a suprise). At present
running `python setup.py build` create the
module inside a build directory. `python
setup.py install` puts it inside the site
packages directory. If you have _ellip_fortran.so
locally you can `import _ellip_fortran` and use the
functions thus:

    import _ellip_fort
    _ellip_fort.ellref(1.3)
    tcor,abrt = _ellip_fort.ellcor('P       ', 2.3, 10, 1.3, 1.9)

The file `elcordir.tbl` must be in the same
directory.

Several tasks remain:

* Add init.py

* Move the fortran open statment to somewhere
  where we don't need to open the file every
  call - we should probably close the file too.

* create a python module to wrap the Fortran.
  This ought to convert degrees to radians,
  have a function to calculate delta (from
  pairs of lat and long) and minimise the
  need to reopen the fortran file.

* add a utility (python, install, whatever)
  to build the elcordir.tbl from
  ELCOR.dat - deal with where to put this.

* Add some tests - to check we get the expected
  correction.