tomocorr: Rewrite Fortran; add Python interface
tomo_predict.f90 contains a Fortran module which makes available
the routine `predict()' for computing the difference in travel
time, given a set of latitude, longitude and depth points, between
a 1D model (held in a file) and a 3D model (held in a file).

By temporarily making everything in tomo_predict.f90 public, allow the
building of the Python interface with f2py.  Type `make` to build it.

To use from Python:

    >>> import tomo_predict
    >>> file1d = "some_1d_file"; file3d = "some_3d_file"
    >>> dt = tomo_predict.tomo_predict.predict(file1d, file3d, lat,
                                               lon, dep)

This needs a working Fortran compiler and f2py.