See how we can link taup with GeographicLib to do the maths
In order to move from the JVM to the Python implementation
of taup for the tomographic correction, we need to be able
to do something like:

    (lon, lat, dep, distance) = path(TauPy_model, evtlon,
                                     evtlat, evtdep,
                                     stalon, stalat)

It turns out that there is now a Python module that makes
this possible. This just shows the approach. The code
probably needs to be inside tomocor.py.