Fix tomocorr2.py and test
test_tomocorr.sh gives:

Original version    Rewritten version
           0.690                0.689

Using the Java taup tools to give the raypath
and the origional and Andy's restructured
Fortran.

The restructured Fortran can be turned into a
python extension module, and we have wrapped
the python taup implementation in Obspy. Using
this we get:

python tomocorr2.py
PcP:  [0.6932814121246338]

(having fixed the arguments. The difference is
probably due to a different number of points on
the taup raypath. But now we have something we
can use for the correction.